Moving utility methods to ParsingUtilities
